+++ src/bugsnag/client/api/Project.ts
@@ -521,10 +521,11 @@ export class ProjectAPI extends BaseAPI {
     );
     //filter by name or ID if provided
     if (collaboratorNameOrId) {
+      const filterValue = collaboratorNameOrId.toLowerCase();
       collaborators.body = collaborators.body.filter(
         (c) =>
           c.id === collaboratorNameOrId ||
-          c.name?.toLowerCase().includes(collaboratorNameOrId.toLowerCase()),
+          c.name?.toLowerCase().includes(filterValue),
       );
     }
     return collaborators;